added dpc read-write, cleaned up base classes
