+++ py4DSTEM/process/phase/__init__.py
@@ -1,10 +1,22 @@
 _emd_hook = True
 
 from py4DSTEM.process.phase.iterative_dpc import DPCReconstruction
-from py4DSTEM.process.phase.iterative_mixedstate_ptychography import MixedStatePtychographicReconstruction
-from py4DSTEM.process.phase.iterative_multislice_ptychography import MultislicePtychographicReconstruction
-from py4DSTEM.process.phase.iterative_overlap_magnetic_tomography import OverlapMagneticTomographicReconstruction
-from py4DSTEM.process.phase.iterative_overlap_tomography import OverlapTomographicReconstruction
+from py4DSTEM.process.phase.iterative_mixedstate_ptychography import (
+    MixedstatePtychographicReconstruction,
+)
+from py4DSTEM.process.phase.iterative_multislice_ptychography import (
+    MultislicePtychographicReconstruction,
+)
+from py4DSTEM.process.phase.iterative_overlap_magnetic_tomography import (
+    OverlapMagneticTomographicReconstruction,
+)
+from py4DSTEM.process.phase.iterative_overlap_tomography import (
+    OverlapTomographicReconstruction,
+)
 from py4DSTEM.process.phase.iterative_parallax import ParallaxReconstruction
-from py4DSTEM.process.phase.iterative_ptychography import PtychographicReconstruction
-from py4DSTEM.process.phase.iterative_simultaneous_ptychography import SimultaneousPtychographicReconstruction
+from py4DSTEM.process.phase.iterative_simultaneous_ptychography import (
+    SimultaneousPtychographicReconstruction,
+)
+from py4DSTEM.process.phase.iterative_singleslice_ptychography import (
+    SingleslicePtychographicReconstruction,
+)
