+++ py4DSTEM/process/phase/iterative_mixedstate_ptychography.py
@@ -20,7 +20,7 @@ except ImportError:
 
 from emdfile import Custom, tqdmnd
 from py4DSTEM import DataCube
-from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
+from py4DSTEM.process.phase.iterative_base_class import PtychographicReconstruction
 from py4DSTEM.process.phase.utils import (
     ComplexProbe,
     fft_shift,
@@ -34,7 +34,7 @@ from py4DSTEM.process.utils import get_CoM, get_shifted_ar
 warnings.simplefilter(action="always", category=UserWarning)
 
 
-class MixedStatePtychographicReconstruction(PhaseReconstruction):
+class MixedstatePtychographicReconstruction(PtychographicReconstruction):
     """
     Mixed-State Ptychographic Reconstruction Class.
