+++ py4DSTEM/process/phase/iterative_overlap_magnetic_tomography.py
@@ -21,7 +21,7 @@ except ImportError:
 
 from emdfile import Custom, tqdmnd
 from py4DSTEM import DataCube
-from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
+from py4DSTEM.process.phase.iterative_base_class import PtychographicReconstruction
 from py4DSTEM.process.phase.utils import (
     ComplexProbe,
     fft_shift,
@@ -36,7 +36,7 @@ from py4DSTEM.process.utils import electron_wavelength_angstrom, get_CoM, get_sh
 warnings.simplefilter(action="always", category=UserWarning)
 
 
-class OverlapMagneticTomographicReconstruction(PhaseReconstruction):
+class OverlapMagneticTomographicReconstruction(PtychographicReconstruction):
     """
     Overlap Magnetic Tomographic Reconstruction Class.
