+++ py4DSTEM/process/phase/iterative_overlap_tomography.py
@@ -21,8 +21,8 @@ except ImportError:
 
 from emdfile import Custom, tqdmnd
 from py4DSTEM import DataCube
-from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
-from py4DSTEM.process.phase.utils import (  # fourier_rotate_real_volume,
+from py4DSTEM.process.phase.iterative_base_class import PtychographicReconstruction
+from py4DSTEM.process.phase.utils import (
     ComplexProbe,
     fft_shift,
     generate_batches,
@@ -35,7 +35,7 @@ from py4DSTEM.process.utils import electron_wavelength_angstrom, get_CoM, get_sh
 warnings.simplefilter(action="always", category=UserWarning)
 
 
-class OverlapTomographicReconstruction(PhaseReconstruction):
+class OverlapTomographicReconstruction(PtychographicReconstruction):
     """
     Overlap Tomographic Reconstruction Class.
