+++ py4DSTEM/process/phase/iterative_singleslice_ptychography.py
@@ -19,7 +19,7 @@ except ImportError:
 
 from emdfile import Custom, tqdmnd
 from py4DSTEM.classes import DataCube
-from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
+from py4DSTEM.process.phase.iterative_base_class import PtychographicReconstruction
 from py4DSTEM.process.phase.utils import (
     ComplexProbe,
     fft_shift,
@@ -32,7 +32,7 @@ from py4DSTEM.process.utils import get_CoM, get_shifted_ar
 warnings.simplefilter(action="always", category=UserWarning)
 
 
-class PtychographicReconstruction(PhaseReconstruction):
+class SingleslicePtychographicReconstruction(PtychographicReconstruction):
     """
     Iterative Ptychographic Reconstruction Class.
